Replace ProjectRegistry with new Registry
