+++ project-registry/config/fantom-testnet.json
@@ -1,5 +1,5 @@
 {
   "network": "fantom-testnet",
   "startBlock": 11194674,
-  "projectRegistryAddress": "0x984749e408FF0446d8ADaf20E293F2F299396631"
+  "registryAddress": ""
 }
