+++ project-registry/config/fantom.json
@@ -1,5 +1,5 @@
 {
   "network": "fantom",
   "startBlock": 50029203,
-  "projectRegistryAddress": "0x8e1bD5Da87C14dd8e08F7ecc2aBf9D1d558ea174"
+  "registryAddress": ""
 }
