+++ project-registry/config/mainnet.json
@@ -1,5 +1,5 @@
 {
   "network": "mainnet",
   "startBlock": 16071515,
-  "projectRegistryAddress": "0x03506eD3f57892C85DB20C36846e9c808aFe9ef4"
+  "registryAddress": ""
 }
