+++ project-registry/config/optimism.json
@@ -1,5 +1,5 @@
 {
   "network": "optimism",
   "startBlock": 18864952,
-  "projectRegistryAddress": "0x8e1bD5Da87C14dd8e08F7ecc2aBf9D1d558ea174"
+  "registryAddress": ""
 }
