+++ project-registry/package.json
@@ -20,16 +20,16 @@
     "prepare:fantom-testnet": "mustache config/fantom-testnet.json subgraph.template.yaml > subgraph.yaml",
     "codegen": "graph codegen",
     "build": "graph build",
-    "deploy": "graph deploy --node https://api.thegraph.com/deploy/ thelostone-mc/program-factory",
+    "deploy-hs": "graph deploy --product hosted-service lmcorbalan/gc-registry",
     "create-local": "graph create --node http://localhost:8020/ thelostone-mc/program-factory",
     "remove-local": "graph remove --node http://localhost:8020/ thelostone-mc/program-factory",
     "deploy-local": "graph deploy --node http://localhost:8020/ --ipfs http://localhost:5001 thelostone-mc/program-factory",
     "test": "graph test"
   },
   "dependencies": {
-    "@graphprotocol/graph-cli": "0.30.4",
-    "@graphprotocol/graph-ts": "0.27.0",
-    "@openzeppelin/subgraphs": "^0.1.8"
+    "@graphprotocol/graph-cli": "0.51.0",
+    "@graphprotocol/graph-ts": "0.31.0",
+    "@openzeppelin/subgraphs": "0.1.8-5"
   },
   "devDependencies": {
     "matchstick-as": "^0.5.0",
