+++ round/package.json
@@ -27,7 +27,8 @@
     "test": "graph test"
   },
   "dependencies": {
-    "@graphprotocol/graph-ts": "0.27.0",
+    "@graphprotocol/graph-cli": "0.51.0",
+    "@graphprotocol/graph-ts": "0.31.0",
     "@openzeppelin/subgraphs": "^0.1.8"
   },
   "devDependencies": {
